split snippet out of models
