+++ collectors/cveorg/tests/test_collectors.py
@@ -6,7 +6,8 @@ from jira.exceptions import JIRAError
 
 from apps.taskman.service import JiraTaskmanQuerier
 from collectors.cveorg.collectors import CVEorgCollector, CVEorgCollectorException
-from osidb.models import Flaw, Snippet
+from osidb.dmodels.snippet import Snippet
+from osidb.models import Flaw
 from osidb.tests.factories import FlawFactory
 
 pytestmark = pytest.mark.integration
