+++ osidb/dmodels/__init__.py
@@ -17,6 +17,8 @@ __all__ = (
     "PsProduct",
     "PsUpdateStream",
     "SpecialConsiderationPackage",
+    # Snippet cannot be added here
+    # as it would make a cycle with Flaw import
     # Tracker cannot be added here
     # as it would make a cycle with Affect and Flaw import
     "UbiPackage",
