+++ osidb/tests/test_snippet.py
@@ -4,7 +4,8 @@ import pytest
 
 from apps.workflows.models import Workflow
 from apps.workflows.workflow import WorkflowFramework, WorkflowModel
-from osidb.models import Flaw, FlawCVSS, FlawReference, Snippet
+from osidb.dmodels.snippet import Snippet
+from osidb.models import Flaw, FlawCVSS, FlawReference
 from osidb.tests.factories import FlawFactory, SnippetFactory
 
 pytestmark = pytest.mark.unit
